backend/requirements.txt
python-multipart>=0.0.6
email-validator>=2.1.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
httpx>=0.25.0
# Database
sqlalchemy>=2.0.23
alembic>=1.13.0
asyncpg>=0.29.0
redis>=5.0.1
# Task Queue
celery>=5.3.4
# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.1.0
pydantic[email]>=2.5.0
# Crawler
requests>=2.31.0
beautifulsoup4>=4.12.2
fake-useragent>=1.4.0
lxml>=5.1.0
# Utils
python-dateutil>=2.8.2
loguru>=0.7.2
aiofiles>=23.2.1
# Payment
# alipay-sdk-python>=4.1.0
# wechatpay-python>=1.0.0
# Dev
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0